+++ packages/ember-data/tests/integration/relationships/has-many-test.js
@@ -90,7 +90,7 @@ module("integration/relationships/has_many - Has-Many Relationships", {
 test("When a hasMany relationship is accessed, the adapter's findMany method should not be called if all the records in the relationship are already loaded", function() {
   expect(0);
 
-  env.adapter.findMany = function() {
+  env.adapter.findMany = function(store, type, ids, snapshots) {
     ok(false, "The adapter's find method should not be called");
   };
 
@@ -106,7 +106,7 @@ test("When a hasMany relationship is accessed, the adapter's findMany method sho
 test("adapter.findMany only gets unique IDs even if duplicate IDs are present in the hasMany relationship", function() {
   expect(2);
 
-  env.adapter.findMany = function(store, type, ids, records) {
+  env.adapter.findMany = function(store, type, ids, snapshots) {
     equal(type, Chapter, 'type passed to adapter.findMany is correct');
     deepEqual(ids, ['2', '3'], 'ids passed to adapter.findMany are unique');
 
@@ -135,18 +135,18 @@ test("A serializer can materialize a hasMany as an opaque token that can be lazi
 
   // When the store asks the adapter for the record with ID 1,
   // provide some fake data.
-  env.adapter.find = function(store, type, id) {
+  env.adapter.find = function(store, type, id, snapshot) {
     equal(type, Post, "find type was Post");
     equal(id, "1", "find id was 1");
 
     return Ember.RSVP.resolve({ id: 1, links: { comments: "/posts/1/comments" } });
   };
 
-  env.adapter.findMany = function() {
+  env.adapter.findMany = function(store, type, ids, snapshots) {
     throw new Error("Adapter's findMany should not be called");
   };
 
-  env.adapter.findHasMany = function(store, record, link, relationship) {
+  env.adapter.findHasMany = function(store, snapshot, link, relationship) {
     equal(link, "/posts/1/comments", "findHasMany link was /posts/1/comments");
     equal(relationship.type.typeKey, "comment", "relationship was passed correctly");
 
@@ -183,7 +183,7 @@ test("Accessing a hasMany backed by a link multiple times triggers only one requ
     post = env.store.push('post', { id: 1, links: { comments: '/posts/1/comments' } });
   });
 
-  env.adapter.findHasMany = function(store, record, link, relationship) {
+  env.adapter.findHasMany = function(store, snapshot, link, relationship) {
     start();
     count++;
     equal(count, 1, "findHasMany has only been called once");
@@ -223,7 +223,7 @@ test("A hasMany backed by a link remains a promise after a record has been added
     message: DS.belongsTo('post', { async: true })
   });
 
-  env.adapter.findHasMany = function(store, record, link, relationship) {
+  env.adapter.findHasMany = function(store, snapshot, link, relationship) {
     return Ember.RSVP.resolve([
       { id: 1, body: "First" },
       { id: 2, body: "Second" }
@@ -253,11 +253,11 @@ test("A hasMany updated link should not remove new children", function() {
     message: DS.belongsTo('post', { async: true })
   });
 
-  env.adapter.findHasMany = function(store, record, link, relationship) {
+  env.adapter.findHasMany = function(store, snapshot, link, relationship) {
     return Ember.RSVP.resolve([]);
   };
 
-  env.adapter.createRecord = function(store, record, link, relationship) {
+  env.adapter.createRecord = function(store, snapshot, link, relationship) {
     return Ember.RSVP.resolve({
       links: {
         comments: '/some/link'
@@ -293,11 +293,11 @@ test("A hasMany updated link should not remove new children when the parent reco
     message: DS.belongsTo('post', { async: true })
   });
 
-  env.adapter.findHasMany = function(store, record, link, relationship) {
+  env.adapter.findHasMany = function(store, snapshot, link, relationship) {
     return Ember.RSVP.resolve([{ id: 5, body: 'hello' }]);
   };
 
-  env.adapter.createRecord = function(store, record, link, relationship) {
+  env.adapter.createRecord = function(store, snapshot, link, relationship) {
     return Ember.RSVP.resolve({
       links: {
         comments: '/some/link'
@@ -330,14 +330,14 @@ test("A hasMany relationship can be reloaded if it was fetched via a link", func
     comments: DS.hasMany('comment', { async: true })
   });
 
-  env.adapter.find = function(store, type, id) {
+  env.adapter.find = function(store, type, id, snapshot) {
     equal(type, Post, "find type was Post");
     equal(id, "1", "find id was 1");
 
     return Ember.RSVP.resolve({ id: 1, links: { comments: "/posts/1/comments" } });
   };
 
-  env.adapter.findHasMany = function(store, record, link, relationship) {
+  env.adapter.findHasMany = function(store, snapshot, link, relationship) {
     equal(relationship.type, Comment, "findHasMany relationship type was Comment");
     equal(relationship.key, 'comments', "findHasMany relationship key was comments");
     equal(link, "/posts/1/comments", "findHasMany link was /posts/1/comments");
@@ -355,7 +355,7 @@ test("A hasMany relationship can be reloaded if it was fetched via a link", func
       equal(comments.get('isLoaded'), true, "comments are loaded");
       equal(comments.get('length'), 2, "comments have 2 length");
 
-      env.adapter.findHasMany = function(store, record, link, relationship) {
+      env.adapter.findHasMany = function(store, snapshot, link, relationship) {
         equal(relationship.type, Comment, "findHasMany relationship type was Comment");
         equal(relationship.key, 'comments', "findHasMany relationship key was comments");
         equal(link, "/posts/1/comments", "findHasMany link was /posts/1/comments");
@@ -379,7 +379,7 @@ test("A sync hasMany relationship can be reloaded if it was fetched via ids", fu
     comments: DS.hasMany('comment')
   });
 
-  env.adapter.find = function(store, type, id) {
+  env.adapter.find = function(store, type, id, snapshot) {
     equal(type, Post, "find type was Post");
     equal(id, "1", "find id was 1");
 
@@ -394,7 +394,7 @@ test("A sync hasMany relationship can be reloaded if it was fetched via ids", fu
       equal(comments.get('isLoaded'), true, "comments are loaded");
       equal(comments.get('length'), 2, "comments have a length of 2");
 
-      env.adapter.findMany = function(store, type, ids, records) {
+      env.adapter.findMany = function(store, type, ids, snapshots) {
         return Ember.RSVP.resolve([
           { id: 1, body: "FirstUpdated" },
           { id: 2, body: "Second" }
@@ -413,14 +413,14 @@ test("A hasMany relationship can be reloaded if it was fetched via ids", functio
     comments: DS.hasMany('comment', { async: true })
   });
 
-  env.adapter.find = function(store, type, id) {
+  env.adapter.find = function(store, type, id, snapshot) {
     equal(type, Post, "find type was Post");
     equal(id, "1", "find id was 1");
 
     return Ember.RSVP.resolve({ id: 1, comments: [1,2] });
   };
 
-  env.adapter.findMany = function(store, type, ids, records) {
+  env.adapter.findMany = function(store, type, ids, snapshots) {
     return Ember.RSVP.resolve([
       { id: 1, body: "First" },
       { id: 2, body: "Second" }
@@ -434,7 +434,7 @@ test("A hasMany relationship can be reloaded if it was fetched via ids", functio
       equal(comments.get('isLoaded'), true, "comments are loaded");
       equal(comments.get('length'), 2, "comments have 2 length");
 
-      env.adapter.findMany = function(store, type, ids, records) {
+      env.adapter.findMany = function(store, type, ids, snapshots) {
         return Ember.RSVP.resolve([
           { id: 1, body: "FirstUpdated" },
           { id: 2, body: "Second" }
@@ -453,14 +453,14 @@ test("A hasMany relationship can be directly reloaded if it was fetched via ids"
     comments: DS.hasMany('comment', { async: true })
   });
 
-  env.adapter.find = function(store, type, id) {
+  env.adapter.find = function(store, type, id, snapshot) {
     equal(type, Post, "find type was Post");
     equal(id, "1", "find id was 1");
 
     return Ember.RSVP.resolve({ id: 1, comments: [1,2] });
   };
 
-  env.adapter.findMany = function(store, type, ids, records) {
+  env.adapter.findMany = function(store, type, ids, snapshots) {
     return Ember.RSVP.resolve([
       { id: 1, body: "FirstUpdated" },
       { id: 2, body: "Second" }
@@ -485,7 +485,7 @@ test("PromiseArray proxies createRecord to its ManyArray once the hasMany is loa
     comments: DS.hasMany('comment', { async: true })
   });
 
-  env.adapter.findHasMany = function(store, record, link, relationship) {
+  env.adapter.findHasMany = function(store, snapshot, link, relationship) {
     return Ember.RSVP.resolve([
       { id: 1, body: "First" },
       { id: 2, body: "Second" }
@@ -516,7 +516,7 @@ test("PromiseArray proxies evented methods to its ManyArray", function() {
     comments: DS.hasMany('comment', { async: true })
   });
 
-  env.adapter.findHasMany = function(store, record, link, relationship) {
+  env.adapter.findHasMany = function(store, snapshot, link, relationship) {
     return Ember.RSVP.resolve([
       { id: 1, body: "First" },
       { id: 2, body: "Second" }
@@ -567,7 +567,7 @@ test("An updated `links` value should invalidate a relationship cache", function
     comments: DS.hasMany('comment', { async: true })
   });
 
-  env.adapter.findHasMany = function(store, record, link, relationship) {
+  env.adapter.findHasMany = function(store, snapshot, link, relationship) {
     equal(relationship.type.typeKey, "comment", "relationship was passed correctly");
 
     if (link === '/first') {
@@ -607,7 +607,7 @@ test("An updated `links` value should invalidate a relationship cache", function
 test("When a polymorphic hasMany relationship is accessed, the adapter's findMany method should not be called if all the records in the relationship are already loaded", function() {
   expect(1);
 
-  env.adapter.findMany = function() {
+  env.adapter.findMany = function(store, type, ids, snapshots) {
     ok(false, "The adapter's find method should not be called");
   };
 
@@ -630,7 +630,7 @@ test("When a polymorphic hasMany relationship is accessed, the store can call mu
     messages: hasMany('message', { polymorphic: true, async: true })
   });
 
-  env.adapter.find = function(store, type) {
+  env.adapter.find = function(store, type, id, snapshot) {
     if (type === Post) {
       return Ember.RSVP.resolve({ id: 1 });
     } else if (type === Comment) {
@@ -921,7 +921,7 @@ test("When a record is saved, its unsaved hasMany records should be kept", funct
 
   var post, comment;
 
-  env.adapter.createRecord = function(store, type, record) {
+  env.adapter.createRecord = function(store, type, snapshot) {
     return Ember.RSVP.resolve({ id: 1 });
   };
 
@@ -944,8 +944,8 @@ test("dual non-async HM <-> BT", function() {
     post: DS.belongsTo('post')
   });
 
-  env.adapter.createRecord = function(store, type, record) {
-    var data = record.serialize();
+  env.adapter.createRecord = function(store, type, snapshot) {
+    var data = snapshot.record.serialize();
     data.id = 2;
     return Ember.RSVP.resolve(data);
   };
@@ -976,11 +976,11 @@ test("When an unloaded record is added to the hasMany, it gets fetched once the 
     comments: DS.hasMany('comment', { async: true })
   });
 
-  env.adapter.findMany = function() {
+  env.adapter.findMany = function(store, type, ids, snapshots) {
     return resolve([{ id: 1, body: 'first' }, { id: 2, body: 'second' }]);
   };
 
-  env.adapter.find = function() {
+  env.adapter.find = function(store, type, id, snapshot) {
     return resolve({ id: 3, body: 'third' });
   };
   var post;